refactor: Post,Update 페이지 로직 파일 분리
