+++ src/assets/SearchIcon.tsx
@@ -8,7 +8,7 @@ const SearchIcon = () => {
       viewBox='0 0 14 14'
       fill='none'
       xmlns='http://www.w3.org/2000/svg'>
-      <g clip-path='url(#clip0_1236_3882)'>
+      <g clipPath='url(#clip0_1236_3882)'>
         <path
           d='M5.92 11.34C8.91338 11.34 11.34 8.91338 11.34 5.92C11.34 2.92662 8.91338 0.5 5.92 0.5C2.92662 0.5 0.5 2.92662 0.5 5.92C0.5 8.91338 2.92662 11.34 5.92 11.34Z'
           stroke={theme.isDark ? theme.colors.white : theme.colors.black}
