+++ src/components/Post/Edit/PostForm/style.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const ButtonWrapper = styled.div`
+  position: sticky;
+  display: flex;
+  justify-content: flex-end;
+  bottom: 0;
+  padding-bottom: 10px;
+  margin-top: 56px;
+`;
