--- src/components/Post/Edit/SubmitButton/type.ts
@@ -1,4 +1,0 @@
-export interface SubmitButtonProps {
-  onSubmit: () => void;
-  message: string;
-}
