+++ src/components/Post/Edit/UserInfo/index.tsx
@@ -1,0 +1,26 @@
+import { UserInfoWrapper } from '@/components/Post/Edit/UserInfo/style'; // UserInfoWrapper는 FormContent에서 import한 것을 그대로 사용
+import { Avatar, Text } from '@/components/common';
+import theme from '@/styles/theme';
+
+interface UserInfoProps {
+  fullName?: string;
+  imageSrc?: string;
+}
+
+const UserInfo = ({ fullName, imageSrc }: UserInfoProps) => {
+  return (
+    <UserInfoWrapper>
+      <Avatar size='middle' alt={fullName || '유저네임'} src={imageSrc} />
+      <Text
+        colorType='grayscale'
+        colorNumber={theme.isDark ? '100' : '500'}
+        fontType='body1'
+        tagType='span'
+        style={{ margin: '0 10px' }}>
+        {fullName || 'user'}
+      </Text>
+    </UserInfoWrapper>
+  );
+};
+
+export default UserInfo;
