+++ src/components/Post/Edit/UserInfo/style.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const UserInfoWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 8px;
+`;
