+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
-
+import { Helmet, HelmetProvider } from 'react-helmet-async';
 import ProtectedRoute from '@/components/common/ProtectedRoute';
 
 import App from '@/App';
