+++ src/pages/PostPage/style.ts
@@ -1,19 +1,9 @@
 import styled from 'styled-components';
 
-export const FormContainer = styled.form`
+export const FormPageContainer = styled.div`
   display: flex;
   width: 100vw;
   max-width: 954px;
   flex-direction: column;
-  align-items: flex-end;
   margin: 80px auto;
-`;
-
-export const ButtonWrapper = styled.div`
-  position: sticky;
-  display: flex;
-  justify-content: flex-end;
-  bottom: 0;
-  padding-bottom: 10px;
-  margin-top: 56px;
 `;
